refactor: better export path
